Renamed task from phpStan to PhpstanCheckDeprecation
