--- src/Task/PhpStan/PhpStanExtensionLoader.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Wunderio\GrumPHP\Task\PhpStan;
-
-use Wunderio\GrumPHP\Task\AbstractExternalExtensionLoader;
-
-/**
- * Class PhpStanExtensionLoader.
- *
- * @package Wunderio\GrumPHP\Task
- */
-class PhpStanExtensionLoader extends AbstractExternalExtensionLoader {}
