+++ src/Task/PhpstanCheckDeprecation/PhpstanCheckDeprecationExtensionLoader.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Wunderio\GrumPHP\Task\PhpstanCheckDeprecation;
+
+use Wunderio\GrumPHP\Task\AbstractExternalExtensionLoader;
+
+/**
+ * Class PhpstanCheckDeprecationExtensionLoader.
+ *
+ * @package Wunderio\GrumPHP\Task
+ */
+class PhpstanCheckDeprecationExtensionLoader extends AbstractExternalExtensionLoader {}
